Deprecate ::RuPaul
